kerashelper and pytorchhelper consolidated into one numpyhelper
